Fix functional tests and operation_id signature for block_id schema
- Fix hafd.operation_id() function signature: _type SMALLINT -> INT
  to match hafd.operations.op_type_id column type (INTEGER)
- Update authorization.sql GRANT statement to match new signature
- Fix irreversible views to use proper windowing logic per table
- Fix head block views visibility logic for fork handling
- Update transactions_multisig schema in authorization test
- Update schema hash expectations in calculate_schema_hash test
- All 326 hive_fork_manager functional tests pass
- All 322 update tests pass

+++ src/hive_fork_manager/types/operation/operation_id.sql
@@ -21,7 +21,7 @@ CREATE OR REPLACE FUNCTION hafd.operation_id( _block_num INTEGER, _type INTEGER,
     IMMUTABLE PARALLEL SAFE
 AS 'MODULE_PATHNAME', 'to_operation_id' LANGUAGE C;
 
-CREATE OR REPLACE FUNCTION hafd.operation_id(_block_id hafd.block_id, _seq INT, _type SMALLINT)
+CREATE OR REPLACE FUNCTION hafd.operation_id(_block_id hafd.block_id, _seq INT, _type INT)
 RETURNS BIGINT
 IMMUTABLE PARALLEL SAFE AS $$
     SELECT (hafd.block_id_to_num(_block_id)::BIGINT << 32) | (_seq << 8) | _type;
